finance/cntk-next-day-trend/requirements.txt
grpcio>=1.14.2
grpcio-tools>=1.14.1
cntk==2.6
pandas_datareader>=0.7.0
